Add support for RHSM Pools
Support rhsm_pool_ids parameter (it was documented but not used).
This parameter helps to specify the subscription pool IDs to consume.

A pool ID may be specified as a string - just the pool ID
(ex. 0123456789abcdef0123456789abcdef), or as a dict with the pool ID as the
key, and a quantity as the value (ex. 0123456789abcdef0123456789abcdef: 2).
If the quantity is provided, it is used to consume multiple entitlements from
a pool (the pool must support this).

Change-Id: I5658fa97e32193045c9da29126e824a884054e60