Rename rhsm_repository module to avoid conflict.
The rhsm_repository is now part of the Ansible installation,
however, the version used inside the redhat_subscription role
contains changes the defaul module from Ansible doesn't contain.

As a consequence, if in the playbook execution, some task makes
use of the rhsm_repository module Ansible will locate it inside
its default module libraries and when we get to the point of invoking
the module from ansible-role-redhat-subscription/tasks/portal.yml
it will find out that the module is already loaded and make use
of the Ansible one, instead of the module stored under library/
rhsm_repository.py. This is a temporary solution until we'll update
into Ansible 2.8 for queens and rocky, then both modules will be
in sync and the one from ansible-role-redhat-subscription/library
could be removed.

Change-Id: Id53ebefff1f992781171a1e760fbf5306020cb56
Closes-Bug: #1833519